refactor: renamings and simplifying mappings
